+++ lib/onlyofficeDocumentManager.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * (c) Copyright Ascensio System SIA 2024.
+ * (c) Copyright Ascensio System SIA 2026.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ class OnlyofficeDocumentManager extends DocumentManager
 
     public function __construct($settingsManager, array $docInfo, $formats = null, $systemLangCode = 'en')
     {
-        $formats = new OnlyofficeFormatsManager;
+        $formats = new OnlyofficeFormatsManager();
         parent::__construct($settingsManager, $formats, $systemLangCode);
         $this->docInfo = $docInfo;
     }
@@ -59,8 +59,14 @@ class OnlyofficeDocumentManager extends DocumentManager
         ];
 
         if (!empty($this->getGroupId())) {
-            $data['groupId'] = $groupId;
+            $data['groupId'] = $this->getGroupId();
         }
+
+        if (isset($this->docInfo['path']) && str_contains($this->docInfo['path'], 'exercises/')) {
+            $data['doctype'] = 'exercise';
+            $data['docPath'] = urlencode($this->docInfo['path']);
+        }
+
         $jwtManager = new OnlyofficeJwtManager($this->settingsManager);
         $hashUrl = $jwtManager->getHash($data);
 
@@ -76,8 +82,6 @@ class OnlyofficeDocumentManager extends DocumentManager
 
     public function getCallbackUrl(string $fileId)
     {
-        $url = '';
-
         $data = [
             'type' => 'track',
             'courseId' => api_get_course_int_id(),
@@ -87,38 +91,55 @@ class OnlyofficeDocumentManager extends DocumentManager
         ];
 
         if (!empty($this->getGroupId())) {
-            $data['groupId'] = $groupId;
+            $data['groupId'] = $this->getGroupId();
+        }
+
+        if (isset($this->docInfo['path']) && str_contains($this->docInfo['path'], 'exercises/')) {
+            $data['doctype'] = 'exercise';
+            $data['docPath'] = urlencode($this->docInfo['path']);
         }
 
         $jwtManager = new OnlyofficeJwtManager($this->settingsManager);
         $hashUrl = $jwtManager->getHash($data);
 
-        return $url.api_get_path(WEB_PLUGIN_PATH).'onlyoffice/callback.php?hash='.$hashUrl;
+        return api_get_path(WEB_PLUGIN_PATH).'onlyoffice/callback.php?hash='.$hashUrl;
     }
 
-    public function getGobackUrl(string $fileId)
+    public function getGobackUrl(string $fileId): string
     {
         if (!empty($this->docInfo)) {
-            return api_get_path(WEB_CODE_PATH).'document/document.php'
-                                                        .'?cidReq='.Security::remove_XSS(api_get_course_id())
-                                                        .'&id_session='.Security::remove_XSS(api_get_session_id())
-                                                        .'&gidReq='.Security::remove_XSS($this->getGroupId())
-                                                        .'&id='.Security::remove_XSS($this->docInfo['parent_id']);
+            if (isset($this->docInfo['path']) && str_contains($this->docInfo['path'], 'exercises/')) {
+                return api_get_path(WEB_CODE_PATH).'exercise/exercise_submit.php'
+                    .'?cidReq='.Security::remove_XSS(api_get_course_id())
+                    .'&id_session='.Security::remove_XSS(api_get_session_id())
+                    .'&gidReq='.Security::remove_XSS($this->getGroupId())
+                    .'&exerciseId='.Security::remove_XSS($this->docInfo['exercise_id']);
+            }
+
+            return self::getUrlToLocation(api_get_course_id(), api_get_session_id(), $this->getGroupId(), $this->docInfo['parent_id'], $this->docInfo['path'] ?? '');
         }
 
         return '';
     }
 
     /**
-     * Return location file in chamilo documents.
+     * Return location file in Chamilo documents or exercises.
      */
-    public static function getUrlToLocation($courseCode, $sessionId, $groupId, $folderId)
+    public static function getUrlToLocation($courseCode, $sessionId, $groupId, $folderId, $filePath = ''): string
     {
+        if (!empty($filePath) && str_contains($filePath, 'exercises/')) {
+            return api_get_path(WEB_CODE_PATH).'exercise/exercise_submit.php'
+                .'?cidReq='.Security::remove_XSS($courseCode)
+                .'&id_session='.Security::remove_XSS($sessionId)
+                .'&gidReq='.Security::remove_XSS($groupId)
+                .'&exerciseId='.Security::remove_XSS($folderId);
+        }
+
         return api_get_path(WEB_CODE_PATH).'document/document.php'
-                                            .'?cidReq='.Security::remove_XSS($courseCode)
-                                            .'&id_session='.Security::remove_XSS($sessionId)
-                                            .'&gidReq='.Security::remove_XSS($groupId)
-                                            .'&id='.Security::remove_XSS($folderId);
+            .'?cidReq='.Security::remove_XSS($courseCode)
+            .'&id_session='.Security::remove_XSS($sessionId)
+            .'&gidReq='.Security::remove_XSS($groupId)
+            .'&id='.Security::remove_XSS($folderId);
     }
 
     public function getCreateUrl(string $fileId)
@@ -181,16 +202,18 @@ class OnlyofficeDocumentManager extends DocumentManager
         int $sessionId,
         int $courseId,
         int $groupId,
-        string $templatePath = ''): array
-    {
+        string $templatePath = '',
+        bool $tryNewFilename = false
+    ): array {
         $courseInfo = api_get_course_info_by_id($courseId);
         $courseCode = $courseInfo['code'];
         $groupInfo = GroupManager::get_group_properties($groupId);
 
-        $fileTitle = Security::remove_XSS($basename).'.'.$fileExt;
-
-        $fileNamePrefix = ChamiloDocumentManager::getDocumentSuffix($courseInfo, $sessionId, $groupId);
-        $fileName = preg_replace('/\.\./', '', $basename).$fileNamePrefix.'.'.$fileExt;
+        $fileNameSuffix = ChamiloDocumentManager::getDocumentSuffix($courseInfo, $sessionId, $groupId);
+        // Try to avoid directories browsing (remove .., slashes and backslashes)
+        $patterns = ['#\.\./#', '#\.\.#', '#/#', '#\\\#'];
+        $replacements = ['', '', '', ''];
+        $sanitizedBasename = preg_replace($patterns, $replacements, $basename).$fileNameSuffix;
 
         if (empty($templatePath)) {
             $templatePath = TemplateManager::getEmptyTemplate($fileExt);
@@ -206,20 +229,30 @@ class OnlyofficeDocumentManager extends DocumentManager
                 $sessionId
             );
             $folderPath = $document_data['absolute_path'];
-            $fileRelatedPath = $fileRelatedPath.substr($document_data['absolute_path_from_document'], 10).'/'.$fileName;
+            $fileRelatedPath = $fileRelatedPath.substr($document_data['absolute_path_from_document'], 10).'/';
         } else {
             $folderPath = api_get_path(SYS_COURSE_PATH).api_get_course_path($courseCode).'/document';
             if (!empty($groupId)) {
                 $folderPath = $folderPath.'/'.$groupInfo['directory'];
                 $fileRelatedPath = $groupInfo['directory'].'/';
             }
-            $fileRelatedPath = $fileRelatedPath.$fileName;
         }
-        $filePath = $folderPath.'/'.$fileName;
+        $filePath = "$folderPath/$sanitizedBasename.$fileExt";
 
-        if (file_exists($filePath)) {
+        if (file_exists($filePath) && !$tryNewFilename) {
             return ['error' => 'fileIsExist'];
         }
+
+        $newFileCounter = 1;
+        $newBasename = $basename;
+        $newSanitizedBasename = $sanitizedBasename;
+        while (file_exists($filePath)) {
+            $newBasename = "$basename ($newFileCounter)";
+            $newSanitizedBasename = "$sanitizedBasename ($newFileCounter)";
+            $filePath = "$folderPath/$newSanitizedBasename.$fileExt";
+            $newFileCounter++;
+        }
+        $fileTitle = Security::remove_XSS($newBasename).'.'.$fileExt;
 
         if ($fp = @fopen($filePath, 'w')) {
             $content = file_get_contents($templatePath);
@@ -230,7 +263,7 @@ class OnlyofficeDocumentManager extends DocumentManager
 
             $documentId = add_document(
                 $courseInfo,
-                $fileRelatedPath,
+                "$fileRelatedPath$newSanitizedBasename.$fileExt",
                 'file',
                 filesize($filePath),
                 $fileTitle,
@@ -255,6 +288,6 @@ class OnlyofficeDocumentManager extends DocumentManager
             }
         }
 
-        return ['documentId' => $documentId];
+        return ['documentId' => $documentId, 'fileTitle' => $fileTitle];
     }
 }