+++ lib/onlyofficeTools.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * (c) Copyright Ascensio System SIA 2024.
+ * (c) Copyright Ascensio System SIA 2026.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -34,7 +34,6 @@ class OnlyofficeTools
 
         $extension = strtolower(pathinfo($document_data['title'], PATHINFO_EXTENSION));
 
-        $canEdit = null !== $documentManager->getFormatInfo($extension) ? $documentManager->getFormatInfo($extension)->isEditable() : false;
         $canView = null !== $documentManager->getFormatInfo($extension) ? $documentManager->getFormatInfo($extension)->isViewable() : false;
 
         $groupId = api_get_group_id();
@@ -47,7 +46,7 @@ class OnlyofficeTools
         $documentId = $document_data['id'];
         $urlToEdit = $urlToEdit.'docId='.$documentId;
 
-        if ($canEdit || $canView) {
+        if ($canView) {
             $tooltip = $plugin->get_lang('openByOnlyoffice');
             if ('pdf' === $extension) {
                 $tooltip = $plugin->get_lang('fillInFormInOnlyoffice');
@@ -152,12 +151,15 @@ class OnlyofficeTools
 
     /**
      * Return path to OnlyOffice viewer for a given file.
-     * @param int $documentId The ID from c_document.iid
-     * @param bool $showHeaders Whether to show Chamilo headers on top of the OnlyOffice frame or not
+     * @param mixed $fileReference File reference (e.g. c_document.iid, file url)
+     * @param bool $showHeaders Show Chamilo headers on top of the OnlyOffice frame
+     * @param int|null $exeId Execution ID
+     * @param int|null $questionId Question ID
+     * @param bool $isReadOnly Read only flag
      *
      * @return string A link to open the OnlyOffice viewer
      */
-    public static function getPathToView(int $documentId, bool $showHeaders = true): string
+    public static function getPathToView($fileReference, bool $showHeaders = true, ?int $exeId = null, ?int $questionId = null, bool $isReadOnly = false): string
     {
         $plugin = OnlyofficePlugin::create();
         $appSettings = new OnlyofficeAppsettings($plugin);
@@ -169,40 +171,61 @@ class OnlyofficeTools
         }
 
         $urlToEdit = api_get_path(WEB_PLUGIN_PATH).'onlyoffice/editor.php';
+        $queryString = $_SERVER['QUERY_STRING'];
+        $isExercise = str_contains($queryString, 'exerciseId=');
 
-        $sessionId = api_get_session_id();
-        $courseInfo = api_get_course_info();
-        $userId = api_get_user_id();
+        if (is_numeric($fileReference)) {
+            $documentId = (int) $fileReference;
+            $courseInfo = api_get_course_info();
+            $sessionId = api_get_session_id();
+            $userId = api_get_user_id();
 
-        $docInfo = DocumentManager::get_document_data_by_id($documentId, $courseInfo['code'], false, $sessionId);
+            $docInfo = DocumentManager::get_document_data_by_id($documentId, $courseInfo['code'], false, $sessionId);
+            if (!$docInfo) {
+                return '';
+            }
 
-        $extension = strtolower(pathinfo($docInfo['path'], PATHINFO_EXTENSION));
-        $canView = null !== $documentManager->getFormatInfo($extension) ? $documentManager->getFormatInfo($extension)->isViewable() : false;
+            $extension = strtolower(pathinfo($docInfo['path'], PATHINFO_EXTENSION));
+            $canView = null !== $documentManager->getFormatInfo($extension) ? $documentManager->getFormatInfo($extension)->isViewable() : false;
 
-        $isGroupAccess = false;
-        $groupId = api_get_group_id();
-        if (!empty($groupId)) {
-            $groupProperties = GroupManager::get_group_properties($groupId);
-            $docInfoGroup = api_get_item_property_info(api_get_course_int_id(), 'document', $documentId, $sessionId);
-            $isGroupAccess = GroupManager::allowUploadEditDocument($userId, $courseInfo['code'], $groupProperties, $docInfoGroup);
+            $isGroupAccess = false;
+            $groupId = api_get_group_id();
+            if (!empty($groupId)) {
+                $groupProperties = GroupManager::get_group_properties($groupId);
+                $docInfoGroup = api_get_item_property_info(api_get_course_int_id(), 'document', $documentId, $sessionId);
+                $isGroupAccess = GroupManager::allowUploadEditDocument($userId, $courseInfo['code'], $groupProperties, $docInfoGroup);
+            }
 
-            $urlToEdit = $urlToEdit.'?groupId='.$groupId.'&';
+            $isMyDir = DocumentManager::is_my_shared_folder($userId, $docInfo['absolute_parent_path'], $sessionId);
+            $accessRights = $isMyDir || $isGroupAccess;
+
+            $urlToEdit .= '?'.api_get_cidreq().'&docId='.$documentId;
+            if (false === $showHeaders) {
+                $urlToEdit .= '&nh=1';
+            }
+
+            if ($canView && !$accessRights) {
+                return $urlToEdit;
+            }
         } else {
-            $urlToEdit = $urlToEdit.'?';
-        }
-        error_log(__LINE__.' '.$urlToEdit);
+            $urlToEdit .= '?'.$queryString.'&doc='.urlencode($fileReference);
+            if ($isExercise) {
+                $urlToEdit .= '&type=exercise';
+                if ($exeId) {
+                    $urlToEdit .= '&exeId='.$exeId;
+                }
 
-        $isMyDir = DocumentManager::is_my_shared_folder($userId, $docInfo['absolute_parent_path'], $sessionId);
+                if ($questionId) {
+                    $urlToEdit .= '&questionId='.$questionId;
+                }
+            }
+            if (false === $showHeaders) {
+                $urlToEdit .= '&nh=1';
+            }
 
-        $accessRights = $isMyDir || $isGroupAccess;
-
-        $urlToEdit = $urlToEdit.'docId='.$documentId;
-        if (false === $showHeaders) {
-            $urlToEdit .= '&nh=1';
-        }
-
-        if ($canView && !$accessRights) {
-            error_log(__LINE__.' '.$urlToEdit);
+            if (true === $isReadOnly) {
+                $urlToEdit .= '&readOnly=1';
+            }
 
             return $urlToEdit;
         }